Move all related components to `storage/registry_contract`
