+++ core/node/consensus/src/storage/registry_contract/mod.rs
@@ -1,0 +1,7 @@
+mod abi;
+pub(crate) mod committee_extractor;
+#[cfg(test)]
+mod testonly;
+#[cfg(test)]
+mod tests;
+pub(crate) mod vm_reader;
